refactor: extract ForbiddenException class
Extract `ForbiddenException` from `AuthorizationCheckBehavior`.

(cherry picked from commit 1f2d4c2302825d28d7f0c701ea1fde4dc01c8da6)
